requirements.txt
aiofiles==0.5.0
aiomysql==0.0.20
argon2-cffi==20.1.0
cffi==1.14.3
cryptography==3.1.1
msgpack==1.0.0
mysqlclient==2.0.1
pycparser==2.20
PyMySQL==0.9.2
PyNaCl==1.4.0
PySocks==1.7.1
six==1.15.0
tornado==6.0.4
